sap_llm_plant_backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
loguru
langchain
openai
